Move methods from ClientPlayerEntityUtils to PlayerUtils
